RF-908: Move files into own dir. Move styles out.
